% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chlorofluoroR-package.R
\docType{data}
\name{CF_demodata}
\alias{CF_demodata}
\title{CF_demodata}
\usage{
data(CF_demodata)
}
\value{
Dataframe in global environment
}
\description{
Example data generated from FluorCam FC800-331 and FluorCam 7
}
\details{
Example data generated from FluorCam FC800-331 and FluorCam 7
Software.
}
\examples{
 data("CF_demodata")
}
