% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chlorofluoroR-package.R
\docType{package}
\name{chlorofluoroR}
\alias{chlorofluoroR}
\alias{chlorofluoroR-package}
\title{chlorofluoroR:}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Data processing and plotting of FvFm and PSII.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kjeynescupper.github.io/chlorofluoroR/}
  \item Report bugs at \url{https://github.com/KJeynesCupper/chlorofluoroR/issues}
}

}
\author{
Katie Jeynes-Cupper \email{kejc@illinois.edu},
Maintainer: Katie Jeynes-Cupper \email{kejc@illinois.edu},
}
\keyword{internal}
