% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chlorofluoroR-package.R
\docType{data}
\name{copynumber}
\alias{copynumber}
\title{copynumber}
\usage{
data(copynumber)
}
\value{
Dataframe in global environment
}
\description{
SExample copy number results.
}
\details{
Example copy number results.
}
\examples{
 data("copynumber")
}
