% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chlorofluoroR-package.R
\docType{data}
\name{layout}
\alias{layout}
\title{layout}
\usage{
data(layout)
}
\value{
Dataframe in global environment
}
\description{
Example plate layout
}
\details{
Example plate layout
}
\examples{
 data("layout")
}
