% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chlorofluoroR-package.R
\docType{data}
\name{selection_results}
\alias{selection_results}
\title{selection_results}
\usage{
data(selection_results)
}
\value{
Dataframe in global environment
}
\description{
Example selection results with BASTA
}
\details{
Example selection results with BASTA
}
\examples{
 data("selection_results")
}
