% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chlorofluoroR-package.R
\docType{data}
\name{step_one}
\alias{step_one}
\title{step_one}
\usage{
data(step_one)
}
\value{
Dataframe in global environment
}
\description{
Example output from computeFvFm_PSII()
}
\details{
Example output from computeFvFm_PSII()
}
\examples{
 data("step_one")
}
