% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_chlorofluoro.R
\name{subset_chlorofluoro}
\alias{subset_chlorofluoro}
\title{ubset for specific combinations}
\usage{
subset_chlorofluoro(
  data,
  output_location,
  copyNumber = 4,
  BASTA = "positive",
  width = 15,
  height = 5
)
}
\arguments{
\item{data}{data.frame or list of data.frame; data output from
computeFvFm_PSII\code{\link[=computeFvFm_PSII]{computeFvFm_PSII()}}}

\item{output_location}{path; location to save output files}

\item{copyNumber}{numeric; copy number to select for}

\item{BASTA}{character; based on chosen value. Default is "positive"}

\item{width}{numeric; pdf width size. deafult is 15}

\item{height}{numeric;pdf height size. deafult is 5}
}
\value{
This outputs a plot and excel document in the same way as
computeFvFm_PSII() function. See output as files "chlorofluoro_plot_2.pdf" and
"chlorofluoro_dataset_2.xlsx"
}
\description{
We want to use this data to help us select for lines to carry to
T2. So, here we will filter for these desired traits. For example, we will
select for those with a copy number value of 4 and BASTA positive
(ie. resistant).
}
\examples{
data("step_one")
output_location <- tempdir()
step_two <- subset_chlorofluoro(step_one, output_location )


}
