% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_chlorofluoro.R
\name{summarise_chlorofluoro}
\alias{summarise_chlorofluoro}
\title{Summarise the FvFm and PSII values for each plant line.}
\usage{
summarise_chlorofluoro(data, output_location, width = 25, height = 25)
}
\arguments{
\item{data}{data.frame or list of data.frame; data output from
computeFvFm_PSII\code{\link[=computeFvFm_PSII]{computeFvFm_PSII()}}}

\item{output_location}{path; location to save output files}

\item{width}{numeric; pdf width size. deafult is 18}

\item{height}{numeric;pdf height size. deafult is 25}
}
\value{
Output is also saved as an excel document
(one sheet per plate and variable). See output as files
"chlorofluoro_plot_3.pdf" and "chlorofluoro_dataset_3.xlsx"
}
\description{
Instead of looking at the distribution of the FvFm or PSII
over multiple light pulses across time, we can calculate a mean. Here we plot
the mean for these variables as a bar plot with error bars representing SD.
As before, the colour is determined by the BASTA selection result.
}
\examples{
data("step_one")
data("selection_results")
output_location <- tempdir()
output_location <- "/Users/kejc/"
step_three <- summarise_chlorofluoro(step_one, output_location)

}
