% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_FvFmPSII.R
\name{summarise_FvFmPSII}
\alias{summarise_FvFmPSII}
\title{Summarise the FvFm and PSII values for each plant line.}
\usage{
summarise_FvFmPSII(
  data,
  copynumber,
  output_location,
  colour_palette = NULL,
  width = 25,
  height = 25
)
}
\arguments{
\item{data}{data.frame or list of data.frame; data output from
computeFvFm_PSII\code{\link[=computeFvFm_PSII]{computeFvFm_PSII()}}}

\item{copynumber}{data.frame; copy number results.}

\item{output_location}{path; location to save output files}

\item{colour_palette}{character; colour palette for ggplot. Default is NULL,
when used plots use packages default green palette (n=19).}

\item{width}{numeric; pdf width size. deafult is 18}

\item{height}{numeric;pdf height size. deafult is 25}
}
\value{
Output is also saved as an excel document
(one sheet per plate and variable). See output as files
"chlorofluoro_plot_3.pdf" and "chlorofluoro_dataset_3.xlsx"
}
\description{
Instead of looking at the distribution of the FvFm or PSII
over multiple light pulses across time, we can calculate a mean. Here we plot
the mean for these variables as a bar plot with error bars representing SD.
As before, the colour is determined by the BASTA selection result.
}
\examples{
data("step_one")
data("selection_results")
data("copynumber")
output_location <- tempdir()
step_three <- summarise_FvFmPSII(step_one, copynumber, output_location)

}
