% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFvFm_PSII.R
\name{computeFvFm_PSII}
\alias{computeFvFm_PSII}
\title{Compute distribution of FvFm and PSII.}
\usage{
computeFvFm_PSII(
  data,
  plate_names,
  layout,
  smartsheet,
  copynumber,
  output_location,
  width = 20,
  height = 21
)
}
\arguments{
\item{data}{data.frame or list of data.frame; from FluorCam software
(frames-numeric)}

\item{plate_names}{character; name of plates.}

\item{layout}{data.frame or list of data.frame; plate layout.}

\item{smartsheet}{data.frame; selection results, ie BASTA.}

\item{copynumber}{data.frame; copy number results.}

\item{output_location}{path; location to save output files}

\item{width}{numeric; pdf width size. deafult is 20.}

\item{height}{numeric;pdf height size. deafult is 21.}
}
\value{
Returns a list of dataframes, where each plates has a dataframe for
FvFm and PSII. Saves an excel file containing all data and a plot for
FvFm and PSII.
}
\description{
Organises the raw data output for 1 or more plates and plots the
distribution.
}
\details{
Takes as input a single dataframe or a list of dataframe. Save data as
chlorofluoro_dataset_1.xlsx and plots as chlorofluoro_plot_1.pdf in desired
directory.
}
\examples{
data("CF_demodata")
data("layout")
data("selection_results")
data("copynumber")

output_location = tempdir()
step_one <- computeFvFm_PSII(CF_demodata,
plate_names = names(CF_demodata),
layout,
smartsheet = selection_results,
copynumber,
output_location)

}
